pyproject.toml
[tool.poetry]
name = "tap-applovin"
version = "0.0.1"
description = "Singer tap for applovin, built with the Meltano Singer SDK."
readme = "README.md"
authors = ["Ryan Bell <ryanbell3669@gmail.com>"]
keywords = [
    "ELT",
    "applovin",
]
classifiers = [
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
license = "Apache-2.0"

[tool.poetry.dependencies]
python = ">=3.9"
singer-sdk = { version="~=0.43.1", extras = [] }
fs-s3fs = { version = "~=1.1.1", optional = true }
requests = "~=2.32.3"
python-dateutil = "~=2.8.2"

[tool.poetry.group.dev.dependencies]
pytest = ">=8"
singer-sdk = { version="~=0.43.1", extras = ["testing"] }

[tool.poetry.extras]
s3 = ["fs-s3fs"]

[tool.pytest.ini_options]
addopts = '--durations=10'

[tool.mypy]
python_version = "3.12"
warn_unused_configs = true

[tool.ruff]
target-version = "py39"

[tool.ruff.lint]
ignore = [
    "COM812",  # missing-trailing-comma
    "ISC001",  # single-line-implicit-string-concatenation
]
select = ["ALL"]

[tool.ruff.lint.flake8-annotations]
allow-star-arg-any = true

[tool.ruff.lint.pydocstyle]
convention = "google"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
# CLI declaration
tap-applovin = 'tap_applovin.tap:Tapapplovin.cli'
